# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for authorization and access token refreshing",
        "title": "Demo OAuth2.0 repository",
        "contact": {},
        "version": "1.2"
    },
    "host": "localhost:8080",
    "basePath": "/restricted",
    "paths": {
        "/getToken/{id}": {
            "get": {
                "description": "call this endpoint to generate and recieve a token pair (jwt access and refresh token). It will return a new token pair in case of success (you have to provide a GUID in URL path).",
                "produces": [
                    "application/json"
                ],
                "summary": "Get token pair",
                "operationId": "getTokenPair",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    }
                }
            }
        },
        "/refreshToken": {
            "post": {
                "description": "call this endpoint to regenerate and recieve a token pair (jwt access and refresh token). It will return a new token pair in case of success (you have to provide a refreshToken in request body).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Refresh token pair",
                "operationId": "refreshTokenPair",
                "parameters": [
                    {
                        "description": "refresh token object",
                        "name": "refreshToken",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RefreshToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    }
                }
            }
        },
        "/restricted": {
            "get": {
                "description": "You have to provide your jwt token to get access to this endpoint. It will return the \"message\": \"got\" in case of success",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "Restricted endpoint (jwt token needed)",
                "operationId": "restricted",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/delivery.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "delivery.Response": {
            "type": "object",
            "properties": {
                "content": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "models.RefreshToken": {
            "type": "object",
            "properties": {
                "guid": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}